chore: remove unused files and align code convention
