+++ packages/core/src/BackgroundProcessManager/BackgroundProcessManager.ts
@@ -1,5 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+
+import { BackgroundManagerNotOpenError } from './BackgroundManagerNotOpenError';
+import { BackgroundProcessManagerState, JobEntry } from './types';
+
 /**
  * @private For internal Amplify use.
  *
@@ -415,61 +419,3 @@ export class BackgroundProcessManager {
 		this._state = BackgroundProcessManagerState.Open;
 	}
 }
-
-/**
- *
- */
-export class BackgroundManagerNotOpenError extends Error {
-	constructor(message: string) {
-		super(`BackgroundManagerNotOpenError: ${message}`);
-	}
-}
-
-/**
- * All possible states a `BackgroundProcessManager` instance can be in.
- */
-export enum BackgroundProcessManagerState {
-	/**
-	 * Accepting new jobs.
-	 */
-	Open = 'Open',
-
-	/**
-	 * Not accepting new jobs. Waiting for submitted jobs to complete.
-	 */
-	Closing = 'Closing',
-
-	/**
-	 * Not accepting new jobs. All submitted jobs are complete.
-	 */
-	Closed = 'Closed',
-}
-
-/**
- * Completely internal to `BackgroundProcessManager`, and describes the structure of
- * an entry in the jobs registry.
- */
-type JobEntry = {
-	/**
-	 * The underlying promise provided by the job function to wait for.
-	 */
-	promise: Promise<any>;
-
-	/**
-	 * Request the termination of the job.
-	 */
-	terminate: () => void;
-
-	/**
-	 * An object provided by the caller that can be used to identify the description
-	 * of the job, which can otherwise be unclear from the `promise` and
-	 * `terminate` function. The `description` can be a string. (May be extended
-	 * later to also support object refs.)
-	 *
-	 * Useful for troubleshooting why a manager is waiting for long periods of time
-	 * on `close()`.
-	 */
-	description?: string;
-};
-
-const process = new BackgroundProcessManager();
