+++ packages/core/src/Reachability/types.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export type NetworkStatus = {
+	online: boolean;
+};
