+++ packages/core/src/utils/retry/index.ts
@@ -1,0 +1,8 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { NonRetryableError } from './NonRetryableError';
+export { isNonRetryableError } from './isNonRetryableError';
+export { jitteredBackoff } from './jitteredBackoff';
+export { jitteredExponentialRetry } from './jitteredExponentialRetry';
+export { retry } from './retry';
