+++ packages/pubsub/src/utils/ReachabilityMonitor/index.native.ts
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { Reachability } from '@aws-amplify/core';
+
+import { Reachability } from '@aws-amplify/core/internals/utils';
 import { default as NetInfo } from '@react-native-community/netinfo';
 
 export const ReachabilityMonitor = () =>
