+++ packages/pubsub/src/utils/ReachabilityMonitor/index.ts
@@ -1,5 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { Reachability } from '@aws-amplify/core';
+import { Reachability } from '@aws-amplify/core/internals/utils';
 
 export const ReachabilityMonitor = () => new Reachability().networkMonitor();
